move logic to lib.rs
